corrigindo exemplos, seperando lib, PSR-4
