+++ src/MediaPost/API/oauth/OAuthException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace MediaPost\API\oauth\OAuthException;
+
+/* Generic exception class
+ */
+class OAuthException extends Exception {
+    // pass
+}
